move APIKeys to Project directory
