+++ apps/studio/components/layouts/ProjectLayout/BuildingState.tsx
@@ -7,7 +7,7 @@ import { Badge, Button } from 'ui'
 import { ClientLibrary } from '@/components/interfaces/Home/ClientLibrary'
 import { ExampleProject } from '@/components/interfaces/Home/ExampleProject'
 import { EXAMPLE_PROJECTS } from '@/components/interfaces/Home/Home.constants'
-import { APIKeys } from '@/components/interfaces/Home/NewProjectPanel/APIKeys'
+import { APIKeys } from '@/components/interfaces/Project/APIKeys'
 import { SupportLink } from '@/components/interfaces/Support/SupportLink'
 import { useInvalidateProjectsInfiniteQuery } from '@/data/projects/org-projects-infinite-query'
 import { useInvalidateProjectDetailsQuery } from '@/data/projects/project-detail-query'
